Extracts code into App.js
