+++ package.json
@@ -8,7 +8,8 @@
   },
   "dependencies": {
     "react": "16.0.0-alpha.6",
-    "react-native": "0.44.0"
+    "react-native": "0.44.0",
+    "react-native-router-flux": "^3.39.1"
   },
   "devDependencies": {
     "babel-jest": "20.0.3",
